Combines FOIA log view files
